BDT/requirements.txt
pandas
numpy
scikit-learn
xgboost
shap
matplotlib
seaborn
requests
python-dotenv
yfinance
ta
pyarrow
